+++ web/WEB-INF/web.xml
@@ -28,6 +28,10 @@
         <servlet-name>UpdateProfileController</servlet-name>
         <servlet-class>controller.UpdateProfileController</servlet-class>
     </servlet>
+    <servlet>
+        <servlet-name>SubmitRequestController</servlet-name>
+        <servlet-class>controller.SubmitRequestController</servlet-class>
+    </servlet>
     <servlet-mapping>
         <servlet-name>ProfileController</servlet-name>
         <url-pattern>/profile</url-pattern>
@@ -128,5 +132,9 @@
     <servlet-mapping>
         <servlet-name>UpdateProfileController</servlet-name>
         <url-pattern>/updateprofile</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>SubmitRequestController</servlet-name>
+        <url-pattern>/submitRequest</url-pattern>
     </servlet-mapping>
 </web-app>
